% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hands.R
\name{calc_defen}
\alias{calc_defen}
\title{Calculate score of hand}
\usage{
calc_defen(
  pai,
  baopai,
  libaopai = "",
  rongpai = "",
  rule = default_rule(),
  zhuangfeng = c("z1", "z2", "z3", "z4"),
  menfeng = c("z2", "z3", "z4", "z1"),
  lizhi = c("none", "lizhi", "double-lizhi"),
  yifa = FALSE,
  qianggang = FALSE,
  lingshang = FALSE,
  haidi = c("none", "haidimoyue", "hedilaoyu"),
  tianhe = c("none", "tianhe", "dihe"),
  changbang = 0L,
  lizhibang = 0L
)
}
\arguments{
\item{pai}{String.
Note that this param is not vectorized.}

\item{baopai}{Character vector. "Dora" indicators.}

\item{libaopai}{Character vector. "Ura-dora" indicators.
Leave empty if there is no libaopai.}

\item{rongpai}{String.
Leave empty if there is no rongpai.}

\item{rule}{List; a rule set.}

\item{zhuangfeng}{String; "ba-kaze" pai.}

\item{menfeng}{String; "ji-kaze" pai.}

\item{lizhi}{String; One of "none", "lizhi", or "double-lizhi".}

\item{yifa}{Logical; flag for "ippatsu".}

\item{qianggang}{Logical; flag for "chankan".}

\item{lingshang}{Logical; flag for "rinshan-kaihou".}

\item{haidi}{String; One of "none", "haidimoyue", or "hedilaoyu".}

\item{tianhe}{String; One of "none", "tianhe", or "dihe".}

\item{changbang}{Integer;
Number of counter sticks that indicates "honba".}

\item{lizhibang}{Integer;
Number of 1,000-point sticks on the table.}
}
\value{
Dataframe.
}
\description{
Calculate score of hand
}
