% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hands.R
\name{calc_xiangting}
\alias{calc_xiangting}
\title{Calculate xiangting number of hands}
\usage{
calc_xiangting(pai)
}
\arguments{
\item{pai}{Character vector.}
}
\value{
Integer vector.
}
\description{
Calculate xiangting number of hands
}
