% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hands.R
\name{collect_tingpai}
\alias{collect_tingpai}
\title{Collect tingpais of hands}
\usage{
collect_tingpai(pai)
}
\arguments{
\item{pai}{Character vector.}
}
\value{
List of character vectors.
}
\description{
Collect all pais that can decrease the xiangting number of hands
if they are drawn.
}
