% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule.R
\name{default_rule}
\alias{default_rule}
\title{Create default rule set}
\usage{
default_rule()
}
\value{
A list.
}
\description{
Create default rule set
}
\seealso{
\url{https://github.com/kobalab/majiang-core/wiki/\%E3\%83\%AB\%E3\%83\%BC\%E3\%83\%AB}
}
