% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{detect-tiles}
\alias{detect-tiles}
\alias{tile2rank}
\alias{tile2suit}
\alias{is_fulou}
\alias{is_tsumogiri}
\alias{is_lizhi}
\alias{is_suit}
\alias{is_yaojiu}
\alias{is_zhongzhang}
\alias{is_suzhi}
\title{Detect specific tiles}
\usage{
tile2rank(x)

tile2suit(x)

is_fulou(x)

is_tsumogiri(x)

is_lizhi(x)

is_suit(x, suit = c("m", "p", "s", "z"))

is_yaojiu(x, negate = FALSE)

is_zhongzhang(x)

is_suzhi(x, suzhi = c("23", "34", "45", "56", "67", "78"))
}
\arguments{
\item{x}{A character vector.}

\item{suit}{A string scalar. Either "m", "p", "s", or "z".}

\item{negate}{A logical scalar.}

\item{suzhi}{A string scalar. Either "23", "34", "45", "56", "67", or "78".}
}
\value{
\itemize{
\item For \code{tile2rank} and \code{tile2suit}: a character vector
(an empty string for "_").
\item For other functions: a logical vector.
}
}
\description{
Detect specific tiles
}
