% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hupai}
\alias{hupai}
\title{List of hupai}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 54 rows and 3 columns.
}
\usage{
hupai
}
\description{
List of hupai
}
\details{
A data frame that contains the list of hupai.

The list includes the following columns:
\itemize{
\item \code{id}: The id of the hupai.
\item \code{en}: The English name of the hupai.
\item \code{jp}: The Japanese name of the hupai.
}
}
\keyword{datasets}
