% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{int2tile}
\alias{int2tile}
\title{Convert integers to tiles}
\usage{
int2tile(x = seq_len(38) - 1, origin = c("zero", "one"))
}
\arguments{
\item{x}{An integer vector.}

\item{origin}{A string scalar. Either "zero" or "one".}
}
\value{
A factor.
}
\description{
Convert integers to tiles
}
\examples{
int2tile(c(0, 1, 25, 37))
}
