% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shikakusphere-package.R
\name{is_valid_pai}
\alias{is_valid_pai}
\title{Check if the pai is a valid tile}
\usage{
is_valid_pai(x)
}
\arguments{
\item{x}{A character vector to be checked.}
}
\value{
A logical vector.
}
\description{
Check if the pai is a valid tile
}
\keyword{internal}
