% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule.R
\name{is_valid_rule}
\alias{is_valid_rule}
\title{Check if a rule is valid}
\usage{
is_valid_rule(rule)
}
\arguments{
\item{rule}{A list.}
}
\value{
A logical scalar.
}
\description{
Check if a rule is valid
}
\keyword{internal}
