% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid_tile}
\alias{is_valid_tile}
\title{Check if each element is a valid tile}
\usage{
is_valid_tile(x)
}
\arguments{
\item{x}{A character vector to be checked.}
}
\value{
A logical vector.
}
\description{
Check if each element is a valid tile
}
