% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lineup}
\alias{lineup}
\title{Line up tiles}
\usage{
lineup(x)
}
\arguments{
\item{x}{A data frame with columns \code{id}, \code{tile}, and \code{n}.}
}
\value{
An object of \code{x$tile} class.
}
\description{
Arrange all tiles from a data frame.
}
\examples{
rand_hands()(3) |>
  paistr() |>
  tidy() |>
  lineup()
}
