% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shikakusphere-package.R
\name{lipai}
\alias{lipai}
\title{Compose hands from character vectors}
\usage{
lipai(x)
}
\arguments{
\item{x}{Character vector or a list of character vectors.}
}
\value{
Character vector.
}
\description{
Compose hands from character vectors
while ignoring invalid pais.
}
\details{
This function accepts less or more than 14 pais per hand,
however, cannot handle 5th pai of each tiles.
If there are any 5th pais, arises an error.
}
\examples{
lipai(list(c("m1", "m2", "m3"), c("p1", "p2", "p3")))
}
