% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjai.R
\name{convert-from-mjai}
\alias{convert-from-mjai}
\alias{mjai_target}
\alias{mjai_conv}
\title{Convert symbols from 'mjai' format}
\usage{
mjai_target(perspective, target)

mjai_conv(type, pai, consumed, target)
}
\arguments{
\item{perspective}{An integer vector
that indicates the plyaer's perspective.}

\item{target}{An integer vector
that represents the target of 'mjai' events.}

\item{type}{A character vector that represents the type of 'mjai' events.}

\item{pai}{A character vector of tiles to be converted
from 'mjai' format into 'cmajiang' format.}

\item{consumed}{A list of character vectors
that represent the 'consumed' tiles in 'mjai' format.}
}
\value{
A character vector.
}
\description{
Converts symbols from 'mjai' format into 'cmajiang' format.
To read 'mjai' format logs, use \code{convlog::read_tenhou6()}.
}
\examples{
if (requireNamespace("dplyr", quietly = TRUE)) {
  data("ranked_game")
  ranked_game[["paifu"]] |>
    dplyr::filter(
      type \%in\% c("tsumo", "dahai", "chi", "pon", "daiminkan", "kakan", "ankan")
    ) |>
    dplyr::mutate(
      target = mjai_target(actor, target),
      pai = trans_tile(pai)
    ) |>
    dplyr::group_by(game_id, round_id, actor) |>
    dplyr::mutate(pai = mjai_conv(type, pai, consumed, target)) |>
    dplyr::ungroup()
}
}
