% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule.R
\name{modify_rule}
\alias{modify_rule}
\title{Modify the default rule set by patch}
\usage{
modify_rule(patch = list())
}
\arguments{
\item{patch}{List.}
}
\value{
List.
}
\description{
Modify the default rule set by patch
}
