% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proceed.R
\name{proceed}
\alias{proceed}
\title{Proceed state of hands}
\usage{
proceed(qipai, zimo, dapai)
}
\arguments{
\item{qipai}{A list of character vectors
where each element represents a player's qipai.}

\item{zimo}{A list of character vectors
where each element represents a players's zimo.}

\item{dapai}{A list of character vectors
where each element represents a players's dapai.}
}
\value{
A character vector that represents proceeded hands.
}
\description{
Proceed state of hands
}
\examples{
if (requireNamespace("dplyr", quietly = TRUE)) {
  qipai <-
    paistr("m123456789p1s1z11") |>
    tidy() |>
    dplyr::mutate(tile = as.character(tile)) |>
    lineup()
  zimo <- list(c("p2", "p3"))
  dapai <- list(c("z1", "z1"))
  proceed(qipai, zimo, dapai)
}
}
