% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand.R
\name{rand_hands}
\alias{rand_hands}
\title{Create a function to randomly generate hands}
\usage{
rand_hands(
  hupai = c("pinghe", "zhuangfeng", "menfeng", "fanpai", "duanyaojiu", "yibeikou",
    "sansetongshun", "yiqitongguan", "hunquandaiyaojiu", "qiduizi", "duiduihu",
    "sananke", "sangangzi", "sansetongke", "hunlaotou", "xiaosanyuan", "hunyise",
    "chunquandaiyaojiu", "erbeikou", "qingyise", "guoshiwushuang", "sianke", "dasanyuan",
    "xiaosixi", "dasixi", "ziyise", "lvyise", "qinglaotou", "sigangzi", "jiulianbaodeng"),
  seed = floor(runif(1, 0, 1) * 1e+05),
  rule = default_rule()
)
}
\arguments{
\item{hupai}{String.}

\item{seed}{Integer.}

\item{rule}{List; a rule set.}
}
\value{
Function inheriting class \code{purrr_function_partial}.
}
\description{
Create a function to randomly generate hands
}
