% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand.R
\name{rand_qipai}
\alias{rand_qipai}
\title{Generate hands at random}
\usage{
rand_qipai(nrow, rule = default_rule(), seed = sample.int(10000, 1))
}
\arguments{
\item{nrow}{An integer scalar.}

\item{rule}{A list; a rule set. Defaults to \code{default_rule()}.}

\item{seed}{An integer scalar.
Used for initializing the random number generator.}
}
\value{
A tibble that consists of 4 columns of \code{paistr} vector.
}
\description{
Generate hands at random
}
\examples{
rand_qipai(3)
}
