% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ranked_game}
\alias{ranked_game}
\title{Sample data of mahjong game log}
\format{
An object of class \code{list} of length 3.
}
\source{
\url{https://github.com/Equim-chan/mjai-reviewer/blob/master/convlog/tests/testdata/ranked_game.json}
}
\usage{
ranked_game
}
\description{
Sample data of mahjong game log
}
\details{
A named list that contains sample data of mahjong game log.
The data was coverted from 'tenhou.net/6' format
using \code{read_tenhou6()} function
in \href{https://github.com/paithiov909/convlog}{convlog} package.
}
\keyword{datasets}
