% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shikakusphere-package.R
\docType{package}
\name{shikakusphere-package}
\alias{shikakusphere}
\alias{shikakusphere-package}
\title{shikakusphere: Miscellaneous Functions for Japanese Mahjong}
\description{
A collection of miscellaneous functions for Japanese mahjong that wraps C++ sources of 'shanten-number' \url{https://github.com/tomohxx/shanten-number} and 'cmajiang' \url{https://github.com/TadaoYamaoka/cmajiang}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://paithiov909.github.io/shikakusphere/}
  \item Report bugs at \url{https://github.com/paithiov909/shikakusphere/issues}
}

}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

Other contributors:
\itemize{
  \item tomo hxx (shanten-number) [copyright holder]
  \item Tadao Yamaoka (cmajiang) [copyright holder]
  \item Lee Thomason (tinyxml2) [copyright holder]
}

}
\keyword{internal}
