% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paistr.R
\name{skksph-paistr}
\alias{skksph-paistr}
\alias{new_paistr}
\title{Internal function to create a \code{skksph_paistr}}
\usage{
new_paistr(x = character())
}
\arguments{
\item{x}{Character vector.}
}
\value{
An object of class \code{skksph_paistr}.
}
\description{
The user-facing wrapper of this function is \code{paistr()}.
}
\keyword{internal}
