% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{skksph_bingpai_to_table}
\alias{skksph_bingpai_to_table}
\title{Internal function to convert bingpai to table}
\usage{
skksph_bingpai_to_table(pai)
}
\arguments{
\item{pai}{hand}
}
\value{
A list of integer vectors
}
\description{
Internal function to convert bingpai to table
}
\keyword{internal}
