% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{skksph_feat_pai}
\alias{skksph_feat_pai}
\title{Internal function to get pai features}
\usage{
skksph_feat_pai(pai)
}
\arguments{
\item{pai}{hand}
}
\value{
A list of numeric vectors
}
\description{
Internal function to get pai features
}
\keyword{internal}
