% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{skksph_get_defen}
\alias{skksph_get_defen}
\title{Internal function to get defen}
\usage{
skksph_get_defen(
  paistr,
  baopai,
  libaopai,
  list,
  rankPoints,
  hongpai,
  rongpai = "",
  zhuangfeng = 0L,
  menfeng = 1L,
  lizhi = 0L,
  yifa = FALSE,
  qianggang = FALSE,
  lingshang = FALSE,
  haidi = 0L,
  tianhu = 0L,
  changbang = 0L,
  lizhibang = 0L
)
}
\arguments{
\item{paistr}{hand}

\item{baopai}{baopai}

\item{libaopai}{libaopai}

\item{list}{rule set}

\item{rankPoints}{rank points (numeric vector of length 4)}

\item{hongpai}{hongpai (numeric vector of length 3)}

\item{rongpai}{rongpai}

\item{zhuangfeng}{zhuangfeng (0-3)}

\item{menfeng}{menfeng (0-3)}

\item{lizhi}{lizhi (0-2)}

\item{yifa}{flag for "ippatsu"}

\item{qianggang}{flag for "chankan"}

\item{lingshang}{flag for "lingshang"}

\item{haidi}{haidi (0-2)}

\item{tianhu}{tianhu (0-2)}

\item{changbang}{changbang}

\item{lizhibang}{lizhibang}
}
\value{
A data frame
}
\description{
Internal function to get defen
}
\keyword{internal}
