% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{skksph_get_n_fulou}
\alias{skksph_get_n_fulou}
\title{Internal function to get number of fulou mianzi}
\usage{
skksph_get_n_fulou(pai)
}
\arguments{
\item{pai}{hand}
}
\value{
An integer vector
}
\description{
Internal function to get number of fulou mianzi
}
\keyword{internal}
