% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{skksph_get_tingpai}
\alias{skksph_get_tingpai}
\title{Internal function to get tingpai}
\usage{
skksph_get_tingpai(pai, index_s, index_h)
}
\arguments{
\item{pai}{hand}

\item{index_s}{internal data (integer matrix)}

\item{index_h}{internal data (integer matrix)}
}
\value{
A data frame
}
\description{
Internal function to get tingpai
}
\keyword{internal}
