% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{skksph_get_xiangting}
\alias{skksph_get_xiangting}
\title{Internal function to get xiangting number}
\usage{
skksph_get_xiangting(pai, index_s, index_h)
}
\arguments{
\item{pai}{hand}

\item{index_s}{internal data (integer matrix)}

\item{index_h}{internal data (integer matrix)}
}
\value{
A data frame
}
\description{
Internal function to get xiangting number
}
\keyword{internal}
