% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{skksph_hand_to_svg}
\alias{skksph_hand_to_svg}
\title{Internal function to convert hand to SVG}
\usage{
skksph_hand_to_svg(pai)
}
\arguments{
\item{pai}{hand}
}
\value{
SVG strings
}
\description{
Internal function to convert hand to SVG
}
\keyword{internal}
