% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{skksph_lipai_impl}
\alias{skksph_lipai_impl}
\title{Internal function to lipai}
\usage{
skksph_lipai_impl(x)
}
\arguments{
\item{x}{list of characters}
}
\value{
list of hands
}
\description{
Internal function to lipai
}
\keyword{internal}
