% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{skksph_tidy_impl}
\alias{skksph_tidy_impl}
\title{Internal function for \verb{tidy(<skksph_paistr>)}}
\usage{
skksph_tidy_impl(pai)
}
\arguments{
\item{pai}{hand}
}
\value{
list of integers
}
\description{
Internal function for \verb{tidy(<skksph_paistr>)}
}
\keyword{internal}
