% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_xiangting}
\alias{test_xiangting}
\title{Test cases for \code{calc_xiangting()}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4000 rows and 3 columns.
}
\source{
\url{https://github.com/TadaoYamaoka/cmajiang/tree/main/tests_cpp/data}
}
\usage{
test_xiangting
}
\description{
Test cases for \code{calc_xiangting()}
}
\details{
A data frame that contains test cases for \code{calc_xiangting()}.
The dataset was extracted from 4 json files with 1000 entries each.
Contents may vary depending on package version.
}
\keyword{datasets}
