% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testdat_xiangting}
\alias{testdat_xiangting}
\title{Test cases for \code{calc_xiangting()}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 40000 rows and 4 columns.
}
\source{
\url{https://github.com/TadaoYamaoka/cmajiang/tree/main/tests_cpp/data}
}
\usage{
testdat_xiangting
}
\description{
Test cases for \code{calc_xiangting()}
}
\details{
A tibble that contains test cases for \code{calc_xiangting()}.
Contents may vary depending on package version.
}
\keyword{datasets}
