% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tiles}
\alias{tiles}
\title{Lookup table for tiles conversion}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 38 rows and 4 columns.
}
\usage{
tiles
}
\description{
Lookup table for tiles conversion
}
\details{
A tibble that contains the lookup table for tiles conversion.
}
\keyword{datasets}
