% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trans_tile}
\alias{trans_tile}
\title{Translate tiles from one format to another}
\usage{
trans_tile(
  x,
  from = c("mjai", "tenhou_int", "id", "cmajiang"),
  to = c("cmajiang", "mjai", "tenhou_int", "id")
)
}
\arguments{
\item{x}{A vector of tiles to be translated.}

\item{from}{A string scalar. Either "mjai", "tenhou_int", "id", or "cmajiang".}

\item{to}{A string scalar. Either "cmajiang", "mjai", "tenhou_int", or "id".}
}
\value{
\itemize{
\item For \code{to = "cmajiang"} and \code{to = "mjai"}: A character vector.
\item For \code{to = "tenhou_int"}: An integer vector.
}
}
\description{
Translate tiles from one format to another
}
\examples{
trans_tile(c("m0", "p1", "z1", "_"), from = "cmajiang", to = "mjai")
trans_tile(c("5mr", "1p", "E", "?"), from = "mjai", to = "cmajiang")
trans_tile(c(51, 21, 41, 0), from = "tenhou_int", to = "mjai")
}
