% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defen.R
\name{calc_defen}
\alias{calc_defen}
\title{Calculate score of hand}
\usage{
calc_defen(
  pai,
  baopai,
  libaopai = "",
  rongpai = "",
  rule = default_rule(),
  zhuangfeng = c("z1", "z2", "z3", "z4"),
  menfeng = c("z2", "z3", "z4", "z1"),
  lizhi = c("none", "lizhi", "double-lizhi"),
  yifa = FALSE,
  qianggang = FALSE,
  lingshang = FALSE,
  haidi = c("none", "haidimoyue", "hedilaoyu"),
  tianhe = c("none", "tianhe", "dihe"),
  changbang = 0L,
  lizhibang = 0L
)
}
\arguments{
\item{pai}{A string scalar. This param is not vectorized.}

\item{baopai}{A character vector. "Dora" indicators.}

\item{libaopai}{A character vector. "Ura-dora" indicators.
Leave empty if there is no libaopai.}

\item{rongpai}{A string scalar such as \code{"m1="}.
Leave empty if there is no rongpai.}

\item{rule}{A list; a rule set. Defaults to \code{default_rule()}.}

\item{zhuangfeng}{A string scalar; "ba-kaze" tile.}

\item{menfeng}{A string scalar; "ji-kaze" tile.}

\item{lizhi}{A string scalar. Either "none", "lizhi", or "double-lizhi".}

\item{yifa}{A logical scalar; flag for "ippatsu".}

\item{qianggang}{A logical scalar; flag for "chankan".}

\item{lingshang}{A logical scalar; flag for "rinshan-kaihou".}

\item{haidi}{A string scalar. Either "none", "haidimoyue", or "hedilaoyu".}

\item{tianhe}{A string scalar; Either "none", "tianhe", or "dihe".}

\item{changbang}{An integer scalar;
the number of counter sticks that indicates "honba".}

\item{lizhibang}{An integer scalar;
the number of 1,000-point sticks on the table.}
}
\value{
A data frame.
}
\description{
Calculate score of hand
}
\examples{
calc_defen("m345567p234s3378", baopai = "z1", rongpai = "s9=")
}
