% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hands.R
\name{calc_defen}
\alias{calc_defen}
\title{Calculate score of hand}
\usage{
calc_defen(
  pai,
  baopai,
  libaopai,
  rongpai = "",
  rule = default_rule(),
  zhuangfeng = 0L,
  menfeng = 1L,
  lizhi = 0L,
  yifa = FALSE,
  qianggang = FALSE,
  lingshang = FALSE,
  haidi = 0L,
  tianhu = 0L,
  changbang = 0L,
  lizhibang = 0L
)
}
\arguments{
\item{pai}{String.
Note that this param is not vectorized.}

\item{baopai}{Character vector. "Dora" indicators.}

\item{libaopai}{Character vector. "Ura-dora" indicators.}

\item{rongpai}{String. Leave empty if there is no rongpai.}

\item{rule}{List; a rule set.}

\item{zhuangfeng}{Integer (0-3); "ba-kaze" pai.}

\item{menfeng}{Integer (0-3); "ji-kaze" pai.}

\item{lizhi}{Integer (0-2);
0: no lizhi, 1: lizhi, 2: double-lizhi.}

\item{yifa}{Logical; flag for "ippatsu".}

\item{qianggang}{Logical; flag for "chankan".}

\item{lingshang}{Logical; flag for "rinshan-kaihou".}

\item{haidi}{Integer (0-2);
0: normal, 1: haidimoyue, 2: hedilaoyu.}

\item{tianhu}{Integer (0-2);
0: normal, 1: tianhu, 2: dihe.}

\item{changbang}{Integer;
Number of counter sticks that indicates "honba".}

\item{lizhibang}{Integer;
Number of 1,000-point sticks on the table.}
}
\value{
Dataframe.
}
\description{
Calculate score of hand
}
