% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hands.R
\name{calc_xiangting}
\alias{calc_xiangting}
\title{Calculate xiangting number of hands}
\usage{
calc_xiangting(pai)
}
\arguments{
\item{pai}{A character vector.}
}
\value{
A tibble.
}
\description{
Calculate xiangting number of hands
}
\examples{
calc_xiangting(c("m345567p234s3378", "p222345z1234567"))
}
