% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tingpai.R
\name{collect_tingpai}
\alias{collect_tingpai}
\title{Collect tingpais for hands}
\usage{
collect_tingpai(pai)
}
\arguments{
\item{pai}{A character vector.}
}
\value{
A list of character vectors.
}
\description{
Collect tiles that can decrease the xiangting number of hands
if they are drawn.
In case there are no tingpais for the hand, the function warns about it
and returns an empty character vector.
}
\examples{
collect_tingpai(c("m345567p234s3378", "p222345z1234567"))
}
