% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shikakusphere-package.R
\name{hand2img}
\alias{hand2img}
\title{Plot a player's hand as an image}
\usage{
hand2img(pai, width = NULL, height = NULL)
}
\arguments{
\item{pai}{String that represents a player's hand.}

\item{width}{Integer scalar.}

\item{height}{Integer scalar.}
}
\value{
A bitmap image that internally converted by \code{magick::image_read_svg}
is invisibly returned.
}
\description{
Plot a player's hand as an image
}
