% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shikakusphere-package.R
\name{hand2img}
\alias{hand2img}
\title{Plot a player's hand as an image}
\usage{
hand2img(pai, ...)
}
\arguments{
\item{pai}{A character vector.}

\item{...}{Other arguments passed to \code{plot()}.}
}
\description{
This function is a short hand for \code{paistr(pai) |> plot()}
}
\seealso{
\code{\link{paistr}}
}
