% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shikakusphere-package.R
\name{lineup}
\alias{lineup}
\title{Line up tiles}
\usage{
lineup(x)
}
\arguments{
\item{x}{A data frame with columns \code{id}, \code{tile}, and \code{n}.}
}
\value{
Character vector (or factor).
}
\description{
Line up tiles
}
\examples{
rand_hands()(3) |>
  paistr() |>
  tidy() |>
  lineup()
}
