% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lipai}
\alias{lipai}
\title{Compose hands from character vectors}
\usage{
lipai(x)
}
\arguments{
\item{x}{A list of character vectors or a character vector.}
}
\value{
A character vector.
}
\description{
Compose hands from character vectors
while ignoring invalid tiles.
This function can handle any number of tiles in each hand,
but cannot more than 5 identical tiles.
If there are more than 5 identical tiles, arises an error.
}
\examples{
lipai(list(c("m1", "m2", "m3"), c("p1", "p2", "p3")))
}
