% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shikakusphere-package.R
\name{parse_hupai}
\alias{parse_hupai}
\title{Parse chains of hupai ids}
\usage{
parse_hupai(str, lang = c("en", "jp"))
}
\arguments{
\item{str}{A character vector.}

\item{lang}{A string scalar. Either "en" or "jp".}
}
\value{
A list of factors.
}
\description{
Parse comma separated chains of hupai ids
into a list of factors.
}
