% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand.R
\name{rand_hands}
\alias{rand_hands}
\title{Create a function to randomly generate hands}
\usage{
rand_hands(
  hupai = c("pinghe", "zhuangfeng", "menfeng", "fanpai", "duanyaojiu", "yibeikou",
    "sansetongshun", "yiqitongguan", "hunquandaiyaojiu", "qiduizi", "duiduihu",
    "sananke", "sangangzi", "sansetongke", "hunlaotou", "xiaosanyuan", "hunyise",
    "chunquandaiyaojiu", "erbeikou", "qingyise", "guoshiwushuang", "sianke", "dasanyuan",
    "xiaosixi", "dasixi", "ziyise", "lvyise", "qinglaotou", "sigangzi", "jiulianbaodeng"),
  zhuangfeng = c("z1", "z2", "z3", "z4"),
  menfeng = c("z2", "z3", "z4", "z1"),
  rule = default_rule(),
  seed = 1234
)
}
\arguments{
\item{hupai}{String.}

\item{zhuangfeng}{String; "ba-kaze" pai.}

\item{menfeng}{String; "ji-kaze" pai.}

\item{rule}{List; a rule set.}

\item{seed}{Integer.}
}
\value{
Function inheriting class \code{purrr_function_partial}.
}
\description{
Create a function to randomly generate hands
}
