% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-json.R
\name{read_tenhou6}
\alias{read_tenhou6}
\title{Read 'tenhou.net/6' format mahjong log}
\usage{
read_tenhou6(json)
}
\arguments{
\item{json}{paths to the JSON files.}
}
\value{
A list that contains the following elements:
\itemize{
\item \code{meta}: a list that contains \code{title}, \code{name} and \code{rule} string of the games.
\item \code{paifu}: a tibble that contains the log of the games.
\item \code{info}: a tibble that contains the other information of the games.
}
}
\description{
Read 'tenhou.net/6' format mahjong log
}
